add fRoomTable, fDelRoom, fAddTable, fSwaptable, function manage Table and Room, fix fMain,  fOrder

+++ MilkTeaHouseProject/DTO/Menu.cs
@@ -12,16 +12,16 @@ namespace MilkTeaHouseProject.DTO
         private int idBill;
         private int idDrink;
         private string drinkName;
-        private int price;
+        private long price;
         private int count;
 
         public string DrinkName { get => drinkName; set => drinkName = value; }
-        public int Price { get => price; set => price = value; }
+        public long Price { get => price; set => price = value; }
         public int Count { get => count; set => count = value; }
         public int IdDrink { get => idDrink; set => idDrink = value; }
         public int IdBill { get => idBill; set => idBill = value; }
 
-        public Menu(int idBill, int idDrink, string drinkName, int price, int count)
+        public Menu(int idBill, int idDrink, string drinkName, long price, int count)
         {
             this.idBill = idBill;
             this.idDrink = idDrink;
@@ -35,7 +35,7 @@ namespace MilkTeaHouseProject.DTO
             this.idBill = (int)row["BillID"];
             this.IdDrink = (int)row["IDDrink"];
             this.drinkName = row["Name"].ToString();
-            this.price = (int)row["Price"];
+            this.price = (long)row["Price"];
             this.count = (int)row["Count"];
         }
     }
